Refactor activeNode and activeDisc.
Did have graphics.activeNode referring to a state.node object and
graphics.activeDisc referring to a graphical element.

Now, states.activeNode() fetches the 'on' node for the current state,
and graphics.activeNode refers to the graphical element. There is no
need for an activeDisc variable anymore (that was only tmp anyway).